# docs/api/openapi_v1.yaml
openapi: 3.0.3
info:
  title: BabyShield API
  version: 1.2.0
  description: >
    BabyShield v1 API for recall search and recall detail.
    Headers:
      - X-API-Version (response header)
    Notes:
      - Use **/api/v1/search/advanced** for all search in apps.
      - Aliases accepted: risk_level→severity, product_category→riskCategory, agency→agencies.
      - Fuzzy search using pg_trgm with similarity threshold 0.08
      - Keyword AND logic for precise matching
      - Deterministic sorting: score DESC, recall_date DESC, id ASC
servers:
  - url: https://babyshield.cureviax.ai
    description: Production server
  - url: http://localhost:8000
    description: Local development
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /api/v1/account:
    delete:
      tags: [account]
      summary: Delete current account (requires recent auth)
      operationId: deleteAccount
      security:
        - bearerAuth: []
      responses:
        "204":
          description: Account deleted successfully. Subsequent calls with same token return 401 (Token revoked).
        "401":
          description: Unauthorized or re-authentication required
        "429":
          description: Too many requests
  /api/v1/user/data/delete:
    post:
      tags: [account-legacy]
      summary: Legacy deletion endpoint (deprecated)
      operationId: legacyDelete
      deprecated: true
      description: "Returns non-2xx (observed 400 when body missing). Use DELETE /api/v1/account instead."
      responses:
        "400":
          description: Bad request (observed when body missing)
        "410":
          description: Gone - endpoint removed
...